# Optimize deep statement matching