--- a/semgrep-core/matching/SubAST_generic.ml
+++ b/semgrep-core/matching/SubAST_generic.ml
@@ -32,7 +32,7 @@ module V = Visitor_AST
 
 (*s: function [[SubAST_generic.subexprs_of_expr]] *)
 (* used for deep expression matching *)
-let subexprs_of_expr e =
+let subexprs_of_expr2 e =
   match e with
   | L _
   | Id _ | IdQualified _  | IdSpecial _
@@ -81,6 +81,8 @@ let subexprs_of_expr e =
     -> []
   | DisjExpr _ -> raise Common.Impossible
 (*e: function [[SubAST_generic.subexprs_of_expr]] *)
+let subexprs_of_expr a = Common.profile_code "Sub.subexprs_of_expr" (fun () ->
+      subexprs_of_expr2 a)
 
 (*****************************************************************************)
 (* Statements *)
@@ -208,7 +210,7 @@ let do_visit_with_ref mk_hooks = fun any ->
 (*e: function [[SubAST_generic.do_visit_with_ref]] *)
 
 (*s: function [[SubAST_generic.lambdas_in_expr]] *)
-let lambdas_in_expr e =
+let lambdas_in_expr2 e =
   do_visit_with_ref (fun aref -> { V.default_visitor with
     V.kexpr = (fun (k, _) e ->
       match e with
@@ -217,31 +219,59 @@ let lambdas_in_expr e =
     );
   }) (E e)
 (*e: function [[SubAST_generic.lambdas_in_expr]] *)
+let lambdas_in_expr a = Common.profile_code "Sub.lambdas_in_expr" (fun () ->
+      lambdas_in_expr2 a)
+
+(* opti: using memoization speed things up a bit too
+ * (but again, this is still slow when called many many times).
+ * todo? note that this is not the optimal memoization we can do because
+ * using Hashtbl where the key is a full expression can be slow (hashing
+ * huge expressions still takes some time). It would be better to
+ * return a unique identifier to each expression to remove the hashing cost.
+ *)
+let hmemo = Hashtbl.create 101
+let lambdas_in_expr_memo2 a =
+  Common.memoized hmemo a (fun () -> lambdas_in_expr a)
+let lambdas_in_expr_memo a = Common.profile_code "Sub.lambdas_in_expr_memo"
+ (fun () -> lambdas_in_expr_memo2 a)
 
 (*****************************************************************************)
 (* Really substmts_of_stmts *)
 (*****************************************************************************)
 
 (*s: function [[SubAST_generic.flatten_substmts_of_stmts]] *)
-let flatten_substmts_of_stmts xs =
+let flatten_substmts_of_stmts2 xs =
+  (* opti: using a ref, List.iter, and Common.push instead of a mix of
+   * List.map, List.flatten and @ below speed things up
+   * (but it is still slow when called many many times)
+   *)
+  let res = ref [] in
+
   let rec aux x =
-    let xs = substmts_of_stmt x in
-    (* getting deeply nested lambdas stmts *)
-    let extras =
-       if not !Flag_semgrep.go_really_deeper_stmt
-       then []
-       else
-         let es = subexprs_of_stmt x in
-         let lambdas = es |> List.map lambdas_in_expr |> List.flatten in
-         lambdas |> List.map (fun def -> def.fbody)
-    in
-
-
     (* return the current statement first, and add substmts *)
-    [x] @
-    (extras |> List.map aux |> List.flatten) @
-    (xs |> List.map aux |> List.flatten)
+    Common.push x res;
+
+    (* this can be really slow because lambdas_in_expr() below can be called
+     * a zillion times on big files (see tests/PERF/) if we do the
+     * matching naively in m_stmts_deep.
+     *)
+    (if not !Flag_semgrep.go_really_deeper_stmt
+    then ()
+    else begin
+       let es = subexprs_of_stmt x in
+       (* getting deeply nested lambdas stmts *)
+       let lambdas = es |> List.map lambdas_in_expr_memo |> List.flatten in
+       lambdas |> List.map (fun def -> def.fbody) |> List.iter aux
+    end);
+
+    let xs = substmts_of_stmt x in
+    xs |> List.iter aux;
   in
-  xs |> List.map aux |> List.flatten
+  xs |> List.iter aux;
+  List.rev !res
 (*e: function [[SubAST_generic.flatten_substmts_of_stmts]] *)
+let flatten_substmts_of_stmts a =
+  Common.profile_code "Sub.flatten_substmts_of_stmts" (fun () ->
+      flatten_substmts_of_stmts2 a)
+
 (*e: semgrep/matching/SubAST_generic.ml *)