--- a/semgrep-core/tests/PERF/l100.py
+++ b/semgrep-core/tests/PERF/l100.py
@@ -1,0 +1,100 @@
+# my_first_calculator.py by AceLewis
+# TODO: Make it work for all floating point numbers too
+
+if 3/2 == 1:  # Because Python 2 does not know maths
+    input = raw_input  # Python 2 compatibility
+
+print('Welcome to this calculator!')
+print('It can add, subtract, multiply and divide whole numbers from 0 to 50')
+num1 = int(input('Please choose your first number: '))
+sign = input('What do you want to do? +, -, /, or *: ')
+num2 = int(input('Please choose your second number: '))
+
+if num1 == 0 and sign == '+' and num2 == 0:
+    print("0+0 = 0")
+if num1 == 0 and sign == '+' and num2 == 1:
+    print("0+1 = 1")
+if num1 == 0 and sign == '+' and num2 == 2:
+    print("0+2 = 2")
+if num1 == 0 and sign == '+' and num2 == 3:
+    print("0+3 = 3")
+if num1 == 0 and sign == '+' and num2 == 4:
+    print("0+4 = 4")
+if num1 == 0 and sign == '+' and num2 == 5:
+    print("0+5 = 5")
+if num1 == 0 and sign == '+' and num2 == 6:
+    print("0+6 = 6")
+if num1 == 0 and sign == '+' and num2 == 7:
+    print("0+7 = 7")
+if num1 == 0 and sign == '+' and num2 == 8:
+    print("0+8 = 8")
+if num1 == 0 and sign == '+' and num2 == 9:
+    print("0+9 = 9")
+if num1 == 0 and sign == '+' and num2 == 10:
+    print("0+10 = 10")
+if num1 == 0 and sign == '+' and num2 == 11:
+    print("0+11 = 11")
+if num1 == 0 and sign == '+' and num2 == 12:
+    print("0+12 = 12")
+if num1 == 0 and sign == '+' and num2 == 13:
+    print("0+13 = 13")
+if num1 == 0 and sign == '+' and num2 == 14:
+    print("0+14 = 14")
+if num1 == 0 and sign == '+' and num2 == 15:
+    print("0+15 = 15")
+if num1 == 0 and sign == '+' and num2 == 16:
+    print("0+16 = 16")
+if num1 == 0 and sign == '+' and num2 == 17:
+    print("0+17 = 17")
+if num1 == 0 and sign == '+' and num2 == 18:
+    print("0+18 = 18")
+if num1 == 0 and sign == '+' and num2 == 19:
+    print("0+19 = 19")
+if num1 == 0 and sign == '+' and num2 == 20:
+    print("0+20 = 20")
+if num1 == 0 and sign == '+' and num2 == 21:
+    print("0+21 = 21")
+if num1 == 0 and sign == '+' and num2 == 22:
+    print("0+22 = 22")
+if num1 == 0 and sign == '+' and num2 == 23:
+    print("0+23 = 23")
+if num1 == 0 and sign == '+' and num2 == 24:
+    print("0+24 = 24")
+if num1 == 0 and sign == '+' and num2 == 25:
+    print("0+25 = 25")
+if num1 == 0 and sign == '+' and num2 == 26:
+    print("0+26 = 26")
+if num1 == 0 and sign == '+' and num2 == 27:
+    print("0+27 = 27")
+if num1 == 0 and sign == '+' and num2 == 28:
+    print("0+28 = 28")
+if num1 == 0 and sign == '+' and num2 == 29:
+    print("0+29 = 29")
+if num1 == 0 and sign == '+' and num2 == 30:
+    print("0+30 = 30")
+if num1 == 0 and sign == '+' and num2 == 31:
+    print("0+31 = 31")
+if num1 == 0 and sign == '+' and num2 == 32:
+    print("0+32 = 32")
+if num1 == 0 and sign == '+' and num2 == 33:
+    print("0+33 = 33")
+if num1 == 0 and sign == '+' and num2 == 34:
+    print("0+34 = 34")
+if num1 == 0 and sign == '+' and num2 == 35:
+    print("0+35 = 35")
+if num1 == 0 and sign == '+' and num2 == 36:
+    print("0+36 = 36")
+if num1 == 0 and sign == '+' and num2 == 37:
+    print("0+37 = 37")
+if num1 == 0 and sign == '+' and num2 == 38:
+    print("0+38 = 38")
+if num1 == 0 and sign == '+' and num2 == 39:
+    print("0+39 = 39")
+if num1 == 0 and sign == '+' and num2 == 40:
+    print("0+40 = 40")
+if num1 == 0 and sign == '+' and num2 == 41:
+    print("0+41 = 41")
+if num1 == 0 and sign == '+' and num2 == 42:
+    print("0+42 = 42")
+if num1 == 0 and sign == '+' and num2 == 43:
+    print("0+43 = 43")