… semgrep-core/tests/PERF/ajin.yaml
… semgrep-core/tests/PERF/ellipsis-js.sgrep
… semgrep-core/tests/PERF/ellipsis-python.sgrep
… semgrep-core/tests/PERF/eq.sgrep
… semgrep-core/tests/PERF/l1000.js
… semgrep-core/tests/PERF/l1000.py
… semgrep-core/tests/PERF/l10000.js
… semgrep-core/tests/PERF/l10000.py
… semgrep-core/tests/PERF/my_first_calculator.py
… semgrep-core/tests/PERF/three.js