This should close #827 and #664

The code to handle foo(); ...; bar(); was very naive
and was doing lots of useless work. This fixes that.

Test plan:
time pipenv run semgrep -f ~/semgrep/tests/PERF/ajin.yaml ~/semgrep/tests/PERF/three.js
3.2s

(was 3min before)

+ /home/pad/github/semgrep/semgrep-core/_build/default/bin/Main.exe -profile -lang py -f tests/PERF/ellipsis-python.sgrep tests/PERF/my_first_calculator.py
---------------------
profiling result
---------------------
Main total                               :      1.670 sec          1 count
Parse_python.parse                       :      0.918 sec          1 count
Parse_python.tokens                      :      0.525 sec          2 count
Semgrep.check                            :      0.458 sec          1 count
Parser_python.main                       :      0.277 sec          1 count
Semgrep.match_sts_sts                    :      0.186 sec      41627 count

(was 85sec before)

+ /home/pad/github/semgrep/semgrep-core/_build/default/bin/Main.exe -profile -lang js -f tests/PERF/ellipsis-js.sgrep tests/PERF/three.js
---------------------
profiling result
---------------------
Main total                               :      2.151 sec          1 count
Parse_js.parse                           :      1.236 sec          1 count
Parse_js.tokens                          :      0.398 sec          2 count
Semgrep.check                            :      0.389 sec          1 count
Semgrep.match_sts_sts                    :      0.239 sec      16824 count
Parser_js.module_item                    :      0.192 sec        609 count

(was a lot more before)