@mschwager 2020-05-28
> Test plan:
> time pipenv run semgrep -f ~/semgrep/tests/PERF/ajin.yaml ~/semgrep/tests/PERF/three.js
> 3.2s
> 
> (was 3min before)

This is amazing!

It looks like a lot of perf tests were added, but none confirming existing behavior. There are a lot of larger changes here, are we confident there's no behavioral regressions without additional functionality tests?