@brendongo 2020-05-28
Rebasing on develop after https://github.com/returntocorp/semgrep/pull/854 should fix tests.