@nbrahms 2020-05-28
Just style thoughts from me, feel free to ignore them.