@brendongo 2020-05-28
Pulled in changes from develop to get tests passing