@aryx 2020-05-29
Yes, I think it's ok to have those 81000 lines; it's convenient to test sometimes performance locally without having to use another repo. Also it boosts my github stats!